+++ lib/rack/dedos/filters/country.rb
@@ -23,7 +23,7 @@ module Rack
 
         def allowed?(request, ip)
           if country = maxmind_db&.get(ip)
-            country_code = country.dig('country', 'iso_code').to_sym
+            country_code = @details = country.dig('country', 'iso_code').to_sym
             @countries.include?(country_code) ? @allowed : !@allowed
           else   # not found in database
             true